Refactor & fix for KmlLayer example
