+++ GoogleMapsComponents/Maps/KmlLayer/KmlLayerMetadata.cs
@@ -1,0 +1,9 @@
+﻿namespace GoogleMapsComponents.Maps.KmlLayer;
+
+public class KmlLayerMetadata : KmlFeatureBase
+{
+    /// <summary>
+    /// Whether the layer has any screen overlays.
+    /// </summary>
+    public bool HasScreenOverlay { get; set; }
+}
